+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - **#1427** An issue introduced in the previous release: the serialization to
   LaTeX of some functions (e.g. `\log`) failed.
 - Serialization to MathML of subscripts/superscripts was incorrect in some cases
+- In Chrome, setting the `readonly` attribute on mathfield caused the content of
+  the mathfield to be set to empty.
 
 ## 0.72.0 (2022-04-18)